refactor(api): move sco-account-recovery service
